refactor/qg-80 Переименовал остальные классы и добавил новый sql скрипт
