+++ src/main/kotlin/pro/qyoga/infra/files/FilesConfig.kt
@@ -1,20 +1,20 @@
-package pro.qyoga.infra.images
+package pro.qyoga.infra.files
 
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
 import org.springframework.context.annotation.Import
 import pro.qyoga.infra.db.SdjConfig
-import pro.qyoga.infra.images.internal.FilesServiceImpl
-import pro.qyoga.infra.images.internal.ImagesRepo
+import pro.qyoga.infra.files.internal.FilesRepo
+import pro.qyoga.infra.files.internal.FilesServiceImpl
 
 @Import(SdjConfig::class)
 @Configuration
-class ImagesConfig(
+class FilesConfig(
     private val sdjConfig: SdjConfig
 ) {
 
     @Bean
-    fun imagesRepo() = ImagesRepo(sdjConfig.jdbcAggregateTemplate(), sdjConfig.jdbcConverter())
+    fun imagesRepo() = FilesRepo(sdjConfig.jdbcAggregateTemplate(), sdjConfig.jdbcConverter())
 
     @Bean
     fun imagesService() = FilesServiceImpl(imagesRepo())
