+++ src/main/kotlin/pro/qyoga/infra/files/api/File.kt
@@ -1,4 +1,4 @@
-package pro.qyoga.infra.images.api
+package pro.qyoga.infra.files.api
 
 import org.springframework.data.annotation.CreatedDate
 import org.springframework.data.annotation.Id
