+++ src/main/resources/db/migration/common/V23111401__rename.sql
@@ -1,0 +1,1 @@
+ALTER TABLE images RENAME TO files;
