+++ src/test/kotlin/pro/qyoga/fixture/ImagesObjectMother.kt
@@ -2,7 +2,7 @@ package pro.qyoga.fixture
 
 import pro.qyoga.fixture.data.imageExtensions
 import pro.qyoga.fixture.data.randomFileName
-import pro.qyoga.infra.images.api.File
+import pro.qyoga.infra.files.api.File
 import kotlin.random.Random
