+++ src/test/kotlin/pro/qyoga/infra/test_config/ClientsTestConfig.kt
@@ -4,16 +4,16 @@ import pro.qyoga.app.therapist.TherapistWebAppConfig
 import pro.qyoga.core.clients.ClientsConfig
 import pro.qyoga.core.programs.exercises.ExercisesConfig
 import pro.qyoga.fixture.ClientsBackgrounds
-import pro.qyoga.infra.images.ImagesConfig
+import pro.qyoga.infra.files.FilesConfig
 
 
 object ClientsTestConfig {
 
-    private val imagesConfig by lazy { ImagesConfig(SdjTestConfig.sdjConfig) }
+    private val filesConfig by lazy { FilesConfig(SdjTestConfig.sdjConfig) }
 
     private val clientsConfig by lazy { ClientsConfig(SdjTestConfig.sdjConfig) }
 
-    private val exercisesConfig by lazy { ExercisesConfig(SdjTestConfig.sdjConfig, imagesConfig) }
+    private val exercisesConfig by lazy { ExercisesConfig(SdjTestConfig.sdjConfig, filesConfig) }
 
     private val therapistWebAppConfig by lazy { TherapistWebAppConfig(clientsConfig, exercisesConfig) }
